Removed examples from serve-d, fixed timespec wrong info, added missing function to dispatcher
Former-commit-id: c6831a25bf9ac4d55d90f946ffbb877d80cbddf1
Former-commit-id: 5657e230586b00fd7a6fc50eee3db5db47624690 [formerly 15bbb5dfd6ef36a4156829d020f7fd09d1d13663]
Former-commit-id: 561d3abfce5de3fdbf9fa6c261140970b2719187

+++ modules/util/source/hip/util/time.d
@@ -18,10 +18,11 @@ version(Windows)
 }
 else
 {
+    import core.stdc.config:c_long;
     struct timespec
     {
         int tv_sec; //Seconds
-        int tv_nsec; //Nanoseconds
+        c_long tv_nsec; //Nanoseconds
     }
     extern(C) int clock_gettime(int clock_id, timespec* tm) nothrow;
 }
@@ -36,7 +37,6 @@ private size_t getSystemTime() nothrow
     }
     else
     {
-        import core.stdc.time;
         timespec tm;
         if(clock_gettime(0, &tm))
             return 0;
